Consolidate codebase: extract shared training/test infrastructure, expand tests
Major refactoring to eliminate ~3,200 lines of duplicated code across species:

- Extract shared training logic into environments/shared/train_base.py (~850 lines)
  replacing ~2,800 lines of near-identical code across three train_sb3.py files.
  Each species script is now a ~30-line thin wrapper with a SpeciesConfig dataclass.

- Extract shared test utilities into environments/shared/test_env_base.py,
  replacing ~500 lines of duplicated test_env.py scripts.

- Add environments/shared/constants.py centralizing sensor layout indices,
  VecNormalize defaults, and physics constants.

- Create consolidated notebooks/training.ipynb parameterized by SPECIES variable,
  replacing three nearly-identical training notebooks.

- Expand test coverage from 234 to 321 tests (+87 new tests):
  - test_train_base.py: linear_schedule, _cast_value, _apply_overrides, SpeciesConfig
  - test_constants.py: regression guards for shared constants
  - test_species_integration.py: cross-species gym.make, config-env integration,
    determinism, observation validity, reward consistency, termination
  - reward_test_helpers.py: reusable reward assertion functions
  - Expanded test_trex_rewards.py from 7 to 25 tests (posture, gait, smoothness,
    height, nosedive, curriculum stages, weight effects)
  - Expanded test_brachio_rewards.py from 10 to 20 tests (gait stability,
    foot contacts, height tracking, curriculum stages, weight effects)

- Rename upload_to_gcs() to _upload_to_gcs() (internal helper)

- Fix: brachiosaurus train_curriculum() now correctly passes eval_callback
  to CurriculumCallback (was missing in the original)
